Built and Implemented C++ Enum Handler
Animal Select is more or less finished. The enum CAnimalT now has functionality through a C++ object that is used by the Animal Selector Widget. Additionally, to showcase the use of incrementation, a second Animal Selector Widget was made based on the first one. They are both functionally very similar, but one is a full list of each animal (based on a predefined array) while the other simply cycles through (based on definition of increment/decrement) each animal.