Exposed MapType to Blueprints
This did not go as desired, but it kind of works.
First, I had to convert MapType to a struct since UCLASS cannot be a UPROPERTY, but USTRUCT can. (???)
Second, while MapOption's MapType factory member is now visible, it unfortunately does not behave as if it were MapTypeEnum like it would in C++, and therefor cannot be set. This may be possible to work around, but maybe not. If not, I'll have to really restructure how MapType works.