MapType Successfully Added to Blueprints
Unfortunately, the class-as-enum technique I originally used does not work well with Blueprints, as the enum was inaccessible. To fix, the enum was separated out into EMapType and the remaining functionality was made into a UCLASS with only static functions that can perform EMapType related operations. As such, FMapOption's MapType was converted to EMapType.