Began Implementing Blueprint Maze Widget
For now, Blueprints can interact with MapGenerator, giving it MapOptions and reading them back. As such, a Map should be able to be generated, but the next step is to determine a method through which Blueprints can interpret the contents of Map and build a physical map. Additionally, the current maze setup widget only utilizes one MapOption, but it is intended to eventually allow the "Add Type" and "Remove Type" buttons allow for multiple types to be used.