Finished Maze Control Menu
The Maze Control Menu now successfully sets the Map Options according to the user's input (and small bugs were fixed as found).
It was a pleasant surprise to find that the TArray of MapOptions, as a static variable, already carries from level to level and doesn't need to be saved via the Game Instance to be transferred over.
With this complete, I'm ready to begin designing a way for Map to be interpreted by Blueprints so that a physical map may be generated.